Adds persistence.
Will now load keymatter from disk if it doesn't exist, and create (and write)
it otherwise.

Also has a sleep to address a race between telling the ACME server to accept a
challenge and requesting a cert on the merits of that challenge. (The server
needs time to check DNS; this client simply has a hard-coded 10s wait.)